Counselor/requirement.txt
Django==5.2.0
djangorestframework==3.15.0
django-cors-headers==4.3.1
django-filter==23.3